requirements.txt
discord
Flask
Pillow
python-dotenv
requests
